1. 카테고리 api req, res 내용 추가

+++ server/src/main/java/com/anamensis/server/service/CategoryService.java
@@ -34,7 +34,7 @@ public class CategoryService {
     }
 
     @Transactional
-    public boolean delete(int id) {
+    public boolean delete(long id) {
         int result = categoryMapper.delete(id);
         return result == 1;
     }
